+++ frontend/src/components/courses/index.jsx
@@ -1,6 +1,5 @@
 export { default as CourseList } from 'components/courses/CourseList'
 export { default as CourseSearchbar } from 'components/courses/CourseSearchbar'
-export { default as Filters } from 'components/courses/Filters'
 export { default as PageNo } from 'components/courses/PageNo'
 export { default as CourseItem } from 'components/courses/item'
 export { default as FiltersBody } from 'components/courses/FiltersBody'
